# core-api/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Core API for Tremolo (auth, users/teachers/friends, game settings, score entries, classes/assignments, dashboard charts).",
        "title": "Tremolo Core API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/users/{userId}/general-info": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-info"
                ],
                "summary": "Get general user info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID (must match the authenticated user)",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.GeneralUserInfoDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID parameter",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dtos.GeneralUserInfoDTO": {
            "type": "object",
            "properties": {
                "created_date": {
                    "description": "Format: \"Joined 12 Mar 2024\"",
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "total_duration": {
                    "description": "Format: \"HH:MM:SS\" or human-readable",
                    "type": "string"
                },
                "total_entries": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}